fix generator info
